Moves creating temporary module to ad_hoc_utils
This is a better place to live.

Makes note that one function is deprecated.
